requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf
groq
langchain_community
phidata
sentence-transformers
python-dotenv
